♻️ Feat(#73): 구조체 개선
