+++ ios/issue-tracker-01/issue-tracker-01/Models/Label.swift
@@ -1,0 +1,21 @@
+//
+//  Label.swift
+//  issue-tracker-01
+//
+//  Created by 조호근 on 5/22/24.
+//
+
+import Foundation
+
+struct LabelResponse: Codable {
+    let id: Int
+    let name: String
+    let description: String
+    let color: String
+}
+
+struct LabelRequest: Codable {
+    let name: String
+    let description: String
+    let color: String
+}
