+++ ios/issue-tracker-01/issue-tracker-01/Models/Milestone.swift
@@ -1,0 +1,13 @@
+//
+//  Milestone.swift
+//  issue-tracker-01
+//
+//  Created by 조호근 on 5/22/24.
+//
+
+import Foundation
+
+struct Milestone: Codable {
+    let id: Int
+    let title: String
+}
